+++ crates/maple_cli/src/cmd/filter/mod.rs
@@ -2,13 +2,115 @@ pub mod dynamic;
 
 pub use dynamic::dyn_fuzzy_filter_and_rank as dyn_run;
 
-use std::collections::HashMap;
-
 use anyhow::Result;
-use fuzzy_filter::{fuzzy_filter_and_rank, Algo, ContentFiltering, Source};
-use printer::truncate_long_matched_lines;
-
+use fuzzy_filter::{fuzzy_filter_and_rank, subprocess, Algo, ContentFiltering, Source};
 use icon::{IconPainter, ICON_LEN};
+use printer::truncate_long_matched_lines;
+use std::collections::HashMap;
+use std::path::PathBuf;
+use structopt::StructOpt;
+
+/// Execute the shell command
+#[derive(StructOpt, Debug, Clone)]
+pub struct Filter {
+    /// Initial query string
+    #[structopt(index = 1, short, long)]
+    query: String,
+
+    /// Filter algorithm
+    #[structopt(short, long, possible_values = &Algo::variants(), case_insensitive = true)]
+    algo: Option<Algo>,
+
+    /// Shell command to produce the whole dataset that query is applied on.
+    #[structopt(short, long)]
+    cmd: Option<String>,
+
+    /// Working directory of shell command.
+    #[structopt(short, long)]
+    cmd_dir: Option<String>,
+
+    /// Synchronous filtering, returns after the input stream is complete.
+    #[structopt(short, long)]
+    sync: bool,
+
+    /// Read input from a file instead of stdin, only absolute file path is supported.
+    #[structopt(long = "input", parse(from_os_str))]
+    input: Option<PathBuf>,
+
+    /// Apply the filter on the full line content or parial of it.
+    #[structopt(short, long, possible_values = &ContentFiltering::variants(), case_insensitive = true)]
+    content_filtering: Option<ContentFiltering>,
+}
+
+impl Filter {
+    /// Firstly try building the Source from shell command, then the input file, finally reading the source from stdin.
+    fn generate_source<I: Iterator<Item = String>>(&self) -> Source<I> {
+        if let Some(ref cmd_str) = self.cmd {
+            if let Some(ref dir) = self.cmd_dir {
+                subprocess::Exec::shell(cmd_str).cwd(dir).into()
+            } else {
+                subprocess::Exec::shell(cmd_str).into()
+            }
+        } else {
+            self.input
+                .clone()
+                .map(Into::into)
+                .unwrap_or(Source::<I>::Stdin)
+        }
+    }
+
+    /// Returns the results until the input stream is complete.
+    #[inline]
+    fn sync_run(
+        &self,
+        number: Option<usize>,
+        winwidth: Option<usize>,
+        icon_painter: Option<IconPainter>,
+    ) -> Result<()> {
+        run::<std::iter::Empty<_>>(
+            &self.query,
+            self.generate_source(),
+            self.algo.clone(),
+            number,
+            icon_painter,
+            winwidth,
+        )
+    }
+
+    #[inline]
+    fn dyn_run(
+        &self,
+        number: Option<usize>,
+        winwidth: Option<usize>,
+        icon_painter: Option<IconPainter>,
+    ) -> Result<()> {
+        dyn_run::<std::iter::Empty<_>>(
+            &self.query,
+            self.generate_source(),
+            self.algo.clone(),
+            number,
+            winwidth,
+            icon_painter,
+            self.content_filtering
+                .clone()
+                .unwrap_or(ContentFiltering::Full),
+        )
+    }
+
+    pub fn run(
+        &self,
+        number: Option<usize>,
+        winwidth: Option<usize>,
+        icon_painter: Option<IconPainter>,
+    ) -> Result<()> {
+        if self.sync {
+            self.sync_run(number, winwidth, icon_painter)?;
+        } else {
+            self.dyn_run(number, winwidth, icon_painter)?;
+        }
+        Ok(())
+    }
+}
 
 /// Returns the info of the truncated top items ranked by the filtering score.
 fn process_top_items<T>(