+++ CHANGELOG.md
@@ -5,12 +5,17 @@ CHANGELOG
 
 ### Added
 
+- New provider `:Clap proj_tags` for project-wide tags.([#391](https://github.com/liuchengxu/vim-clap/pull/391))
 - Allow `:Clap files +name-only` to filter the file name only instead of the full file path. Require you have built the Python dynamic module or uses in the cached mode. ([#389](https://github.com/liuchengxu/vim-clap/pull/389))
 
 ### Improved
 
 - List all the autoloaded providers instead of the builtin ones in `:Clap providers`.
 - Handle the icon highlight offset on Python and Rust side.
+
+### Changed
+
+- Now `:Clap tags` will filter the tag name column only, same with `:Clap proj_tags`.
 
 ### Fixed